fixed users skills section
